+++ Android/AwsMobileHub/.gitignore
