+++ Android/AwsMobileHub/app/.gitignore
