+++ Android/AwsMobileHub/app/proguard-rules.pro
