+++ Android/AwsMobileHub/app/src/main/res/mipmap-mdpi/ic_launcher_round.png
