+++ Android/AwsMobileHub/app/src/main/res/values/colors.xml
