+++ Android/AwsMobileHub/app/src/main/res/values/styles.xml
