+++ Android/AwsMobileHub/gradlew
