+++ Android/AwsMobileHub/gradlew.bat
