+++ Android/AwsMobileHub/settings.gradle
