--- Android/app/google-services.json
@@ -1,55 +1,0 @@
-{
-  "project_info": {
-    "project_number": "169897392571",
-    "firebase_url": "https://realtime-db-44b83.firebaseio.com",
-    "project_id": "realtime-db-44b83",
-    "storage_bucket": "realtime-db-44b83.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:169897392571:android:e39bbf6280d9bd0f",
-        "android_client_info": {
-          "package_name": "com.example.motivation.firebaserealtimedatabase"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "169897392571-e178sabh4d7stpofc59l9rqpugbg0sm3.apps.googleusercontent.com",
-          "client_type": 1,
-          "android_info": {
-            "package_name": "com.example.motivation.firebaserealtimedatabase",
-            "certificate_hash": "faeb3460c734fe1e309c35425c4e7ad46e7b4db0"
-          }
-        },
-        {
-          "client_id": "169897392571-qh1e2jig79ts91masnvamjk72m3nhpcs.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyBoL-1MiaGJyx1nCCoQGw6U5lP4eUh3ML8"
-        }
-      ],
-      "services": {
-        "analytics_service": {
-          "status": 1
-        },
-        "appinvite_service": {
-          "status": 2,
-          "other_platform_oauth_client": [
-            {
-              "client_id": "169897392571-qh1e2jig79ts91masnvamjk72m3nhpcs.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        },
-        "ads_service": {
-          "status": 2
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
