--- Android/app/src/main/java/com/example/motivation/firebaserealtimedatabase/Item.java
@@ -1,21 +1,0 @@
-package com.example.motivation.firebaserealtimedatabase;
-
-public class Item {
-
-    private String title;
-
-    public Item() {
-    }
-
-    public Item(String title) {
-        this.title = title;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-}
