--- Android/app/src/main/res/layout/activity_listview.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".ListviewActivity">
-
-    <TextView
-        android:id="@+id/listview_textview"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="LISTVIEW"
-        android:textAlignment="center"
-        android:textColor="#000"
-        android:textSize="30sp" />
-
-    <ListView
-        android:id="@+id/list_listview"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    </ListView>
-
-</LinearLayout>
