--- Android/app/src/main/res/layout/activity_main.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".MainActivity">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <EditText
-            android:id="@+id/main_item_name"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Item" />
-
-        <EditText
-            android:id="@+id/main_item_content"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="content" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-        <Button
-            android:id="@+id/main_input"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="INPUT" />
-
-        <EditText
-            android:id="@+id/main_search"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Find" />
-
-        <Button
-            android:id="@+id/main_output"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="OUTPUT" />
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/main_outputText"
-        android:layout_width="match_parent"
-        android:layout_height="320dp" />
-
-    <Button
-        android:text="Listview"
-        android:id="@+id/main_listviewButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
-
-</LinearLayout>
