+++ Android/gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,2 @@
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
+#Wed Nov 07 22:26:53 KST 2018
 distributionUrl=https\://services.gradle.org/distributions/gradle-4.6-all.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
